v0.0.2

+++ src/moon/watcher.coffee
@@ -12,10 +12,20 @@ path      = require "path"
 glob      = require "glob"
 jade      = require "jade"
 stylus    = require "stylus"
-diff      = require './diff'
+diff      = require "./diff"
+
+stylus    = require "stylus"
+jade      = require "jade"
+
+prettydiff  = require "./prettydiff"
 
 # Import classes
 Logger = require "./logger"
+
+# node < 0.7 compatibility
+unless fs.existsSync
+  fs.exists = path.exists
+  fs.existsSync = path.existsSync
 
 ###
   Watcher
@@ -25,45 +35,45 @@ class Watcher
   # Private variables
   logger = new Logger "watcher"
   defaults =
-    ext: "js,json,css,jade,stylus,coffee",
+    ext: "js,json,css,jade,styl,coffee,png,jpg,gif",
     include: "**/",
     exclude: "node_modules/**,test/**,package.json"
 
   # Temporary directory
   tempPath = temp.path()
-  unless path.existsSync tempPath
+  unless fs.existsSync tempPath
     fs.mkdirSync tempPath
 
   # Remove temporary directory
-  try
+  unless process.platform is "win32"
     process.on "SIGINT", ()->
       logger.debug "Cleaning up..."
       fs.unlinkSync tempPath + "/*"
       fs.rmdirSync tempPath
-
-  catch e
+  else
     # Fallback for Windows
     # Bug in node.js prevents us to do this
     # Ref: https://groups.google.com/group/nodejs/browse_thread/thread/9a6b9214722e526f
     ###
-    tty = require "tty"
-    process.stdin.resume()
-    tty.setRawMode true
-    process.stdin.on "keypress", (char, key) ->
-      if key and key.ctrl and key.name is "c"
-        logger.debug "Cleaning up..."
-        fs.unlinkSync tempPath + "/*"
-        fs.rmdirSync tempPath
-        process.exit 0
-    ###
+    try
+      tty = require "tty"
+      process.stdin.resume()
+      tty.setRawMode true
+      process.stdin.on "keypress", (char, key) ->
+        if key and key.ctrl and key.name is "c"
+          logger.debug "Cleaning up..."
+          fs.unlinkSync tempPath + "/*"
+          fs.rmdirSync tempPath
+          process.exit 0###
 
   # Public variables
   files: []
+  tempFiles: {}
 
   constructor: (cb, opts) ->
     self = @
     dir = process.cwd()
-    unless path.existsSync dir
+    unless fs.existsSync dir
       logger.error "Path does not exist"
       return
     @dir = dir
@@ -92,25 +102,26 @@ class Watcher
         include.push "#{@opts.include[dir]}*.#{@opts.ext[ext]}"
 
     pattern = "{" + include.join(',') + "}"
-    glob pattern, nonegate: true, (err, files) ->
-      include = files
+    await glob pattern, nonegate: true, defer err,files
+    include = files
 
-      for dir of self.opts.exclude
-        exclude.push "**/#{self.opts.exclude[dir]}"
+    for dir of self.opts.exclude
+      exclude.push "**/#{self.opts.exclude[dir]}"
 
-      pattern = "{" + exclude.join(',') + "}"
-      glob pattern, nonegate: true, (err, files) ->
-        for file of files
-          idx = include.indexOf(files[file])
-          if idx > 0
-            include.splice idx, 1
-        self.files = include
-        if typeof cb is "function" then cb()
+    pattern = "{" + exclude.join(',') + "}"
+    await glob pattern, nonegate: true, defer err, files
+    for file of files
+      idx = include.indexOf(files[file])
+      if idx > 0
+        include.splice idx, 1
+    self.files = include
+    if typeof cb is "function" then cb()
 
   tempCopy: (file) ->
-    tempFile = "#{tempPath}/" + crypto.createHash("md5").update(file).digest("hex")
-    require('fs-extra').copyFile file, tempFile, (err) -> {}
-    return path.normalize tempFile
+    unless @tempFiles[file]
+      @tempFiles[file] = path.normalize "#{tempPath}/" + crypto.createHash("md5").update(file).digest("hex")
+    require('fs-extra').copyFile file, @tempFiles[file], (err) -> {}
+    this
 
   watch: (cb) ->
     self = @
@@ -119,10 +130,9 @@ class Watcher
       fs.stat file, (e, prevStats) ->
         throw e if e
         ext = path.extname(file).replace ".", ""
+
         # Create copies of the original for diff patch
-        switch ext
-          when "jade", "css", "stylus"
-            tempFile = self.tempCopy file
+        self.tempCopy file
 
         watcher = fs.watch file, callback = (e, filename) ->
           if e is "rename"
@@ -136,14 +146,87 @@ class Watcher
 
               logger.debug "Detected change in", file
               prevStats = stats
+
+              oldStr = ( fs.readFileSync self.tempFiles[file] ).toString()
+              newStr = ( fs.readFileSync file .toString() ).toString()
+
+              self.tempCopy file
+
               switch ext
-                when "jade", "css", "stylus"
-                  oldStr = fs.readFileSync(tempFile).toString()
-                  newStr = fs.readFileSync(file).toString()
-                  patch = diff.diffLines oldStr, newStr
-                  cb("push", file, patch)
+
+                when "jpg", "png", "gif"
+                  cb "change", 
+                    action: "reloadSingle",
+                    file: file
+
+                when "jade"
+                  ###
+                  oldStr = do jade.compile oldStr, pretty: true 
+                  newStr = do jade.compile newStr, pretty: true
+
+                  #prettydiff.api( source: oldStr, diff: newStr, mode: "diff",  );
+
+                  patch = diff.diffWords oldStr, newStr
+
+                  for i, change of patch
+                    if change.added or change.removed
+                      console.log change
+                  ###
+                  cb "change", 
+                    action: "reload",
+                    file: file
+
+                when "styl"
+
+                  try
+
+                    await
+                      stylus.render oldStr, pretty: true, defer err, oldStr
+                      stylus.render newStr, pretty: true, defer err, newStr
+
+                    patch = diff.diffCss oldStr, newStr
+
+                    changes = []
+                    for change,i in patch
+                      if !change.added and !change.removed
+                        patch.slice(i,0)
+                        continue
+                      val = change.value = change.value.replace(/\n/g,"").trim()
+                      change.selector = val.match(/^(.[^{]+)/)[0].trim()
+                      change.rules = _.compact(val.split("{").slice(1).join("").trim().split(";"))
+                      changes.push change
+
+                    cb "change", 
+                      action: "reloadSingle",
+                      file: file, 
+                      changes: changes
+
+                  catch err
+                    logger.error "Something went wrong:", err
+                    cb "error",
+                      file: file,
+                      error: err
+
+                when "css"
+
+                  patch = diff.diffCss oldStr, newStr
+                  changes = []
+                  for change,i in patch
+                    if !change.added and !change.removed
+                      patch.slice(i,0)
+                      continue
+                    val = change.value = change.value.replace(/\n/g,"").trim()
+                    change.selector = val.match(/^(.[^{]+)/)[0].trim()
+                    change.rules = _.compact(val.split("{").slice(1).join("").trim().split(";"))
+                    changes.push change
+
+                  cb "change", 
+                    action: "reloadSingle",
+                    file: file, 
+                    changes: changes
 
                 when "js", "coffee"
-                  cb("reload", file)
+                  cb "reload",
+                    file: file
 
 module.exports = Watcher
